Datajoint v2 version

+++ djimaging/tables/classifier_v2/rgc_classifier_v2.py
@@ -10,7 +10,7 @@ class ClassifierV2Template(dj.Manual):
     @property
     def definition(self) -> str:
         definition = """
-        classifier_id : int unsigned  # Unique identifier for the classifier
+        classifier_id : int64  # Unique identifier for the classifier
         ---
         classifier_file : varchar(255)
         """
